requirements.txt
python>=3.6
tqdm=4
psutil=5
pytables=3.5
prodigal>=2.6
aragorn>=1.2
infernal>=1.1
mmseqs2=9
networkx=2.3